requirements.txt
altgraph==0.17.4
attrs==23.2.0
auto-py-to-exe==2.43.3
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
CTkMenuBar==0.8
customtkinter==5.2.2
cx_Logging==3.1.0
darkdetect==0.8.0
Eel==0.16.0
Flask==3.0.2
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
lief==0.14.1
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
playwright==1.41.2
pluggy==1.5.0
pycparser==2.22
pyee==11.0.1
pygame==2.5.2
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pynput==1.7.6
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.2.1
pywin32-ctypes==0.2.2
requests==2.31.0
selenium==4.21.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.2
wheel==0.42.0
whichcraft==0.6.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
